Static Factory CommandHandler method should have compatible return type
An exception is now thrown when the return type of a static
@CommandHandler is not assignable to the class that declares it.
