+++ core/src/main/java/org/axonframework/commandhandling/model/inspection/MethodCommandHandlerDefinition.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2016. Axon Framework
+ * Copyright (c) 2010-2017. Axon Framework
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -64,6 +64,10 @@ public class MethodCommandHandlerDefinition implements HandlerEnhancerDefinition
                 commandName = (String) annotationAttributes.get("commandName");
             }
             final boolean factoryMethod = executable instanceof Method && Modifier.isStatic(executable.getModifiers());
+            if (factoryMethod && !executable.getDeclaringClass().isAssignableFrom(((Method)executable).getReturnType())) {
+                throw new AxonConfigurationException("static @CommandHandler methods must declare a return value " +
+                                                             "which is equal to or a subclass of the declaring time");
+            }
             isFactoryHandler = executable instanceof Constructor || factoryMethod;
         }
